pyproject.toml
[project]
name = "ai-movie-studio"
version = "0.1.0"
description = "CLI for long-form AI video orchestration (schemas, render, assemble)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "jsonschema>=4.23.0",
  "python-dotenv>=1.0.0",
  "typer>=0.12.0",
  "edge-tts>=6.1.0",
]

[project.scripts]
studio = "studio.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["studio"]
